refactor(desktop): retire product and platform legacy aliases
Closes #30306
