+++ turbo/apps/desktop/scripts/desktop-build-config.js
@@ -2,9 +2,7 @@ const fs = require("node:fs");
 const path = require("node:path");
 
 const desktopIdentities = require("../src/desktop-identities.json");
-const {
-  resolveDesktopEnvironmentAlias,
-} = require("./desktop-environment-alias");
+const { readDesktopEnvironment } = require("./desktop-environment");
 
 const PRODUCTION_PLATFORM_HOSTNAMES = new Set(["app.vm0.ai", "app.okou.ai"]);
 const RUNTIME_CONFIG_PATH = path.resolve(
@@ -35,13 +33,13 @@ function resolveDesktopBuildConfig(options = {}) {
   const fileConfig = readRuntimeConfig();
   const product = desktopProduct(
     options.product?.trim() ||
-      resolveDesktopEnvironmentAlias("OKOU_DESKTOP_PRODUCT") ||
+      readDesktopEnvironment("OKOU_DESKTOP_PRODUCT") ||
       fileConfig?.product ||
       "zero",
   );
   const platformUrl = new URL(
     options.platformUrl?.trim() ||
-      resolveDesktopEnvironmentAlias("OKOU_DESKTOP_PLATFORM_URL") ||
+      readDesktopEnvironment("OKOU_DESKTOP_PLATFORM_URL") ||
       fileConfig?.platformUrl ||
       desktopIdentities[product].defaultPlatformUrl,
   );
