--- turbo/apps/desktop/scripts/desktop-environment-alias.d.ts
@@ -1,7 +1,0 @@
-export type DesktopEnvironmentAliasKey =
-  | "OKOU_DESKTOP_PLATFORM_URL"
-  | "OKOU_DESKTOP_PRODUCT";
-
-export function resolveDesktopEnvironmentAlias(
-  canonicalKey: DesktopEnvironmentAliasKey,
-): string | undefined;
