+++ turbo/apps/desktop/scripts/desktop-environment.d.ts
@@ -1,0 +1,3 @@
+export function readDesktopEnvironment(
+  key: "OKOU_DESKTOP_PLATFORM_URL" | "OKOU_DESKTOP_PRODUCT",
+): string | undefined;
