+++ turbo/apps/desktop/scripts/desktop-environment.js
@@ -1,0 +1,5 @@
+function readDesktopEnvironment(key) {
+  return process.env[key]?.trim() || undefined;
+}
+
+module.exports = { readDesktopEnvironment };
